+++ scripts/layerdiffuse/layerdiffuse_model.py
@@ -11,10 +11,7 @@ from diffusers.configuration_utils import ConfigMixin, register_to_config
 from diffusers.models.modeling_utils import ModelMixin
 from diffusers.models.autoencoders.vae import DecoderOutput
 from diffusers.models.attention_processor import Attention, AttnProcessor
-try:
-    from diffusers.models.unet_2d_blocks import UNetMidBlock2D, get_down_block, get_up_block
-except Exception:
-    from diffusers.models.unets.unet_2d_blocks import UNetMidBlock2D, get_down_block, get_up_block
+from diffusers.models.unets.unet_2d_blocks import UNetMidBlock2D, get_down_block, get_up_block
 
 
 def zero_module(module):